897996: After selecting the values, the selected values are in disabled
### Bug description
Clearly and concisely describe the problem (this cannot be empty).

### Root cause
Briefly describe the root cause and analysis of the problem. If there is an internal discussion on the forum, provide the link.

### Solution description
Describe the changes made in the code in detail for the reviewers.

### Reason for not identifying earlier
 * [ ] Guidelines not followed. If yes, provide which guideline is not followed.

 * [ ] Guidelines not given. If yes, provide which/who need to address. Tag label `update-guideline-coreteam` or `update-guideline-productteam`. 

 * [ ] If any other reason, provide the details here. 
     
### Areas tested against this fix
Provide details about the areas or combinations that have been tested against this code changes.
* [ ]  Tested against feature matrix. [Feature matrix link](https://syncfusion.sharepoint.com/sites/EJ2ProductOwners/Shared%20Documents/Forms/AllItems.aspx?viewid=ae81c682%2D3d0f%2D462a%2Db8ec%2D7358748d489d&id=%2Fsites%2FEJ2ProductOwners%2FShared%20Documents%2FGeneral%2FFeature%20Matrix%20%2D%20Documents)
* [ ]  NA

### Is it a breaking issue?
* [ ]  Yes, Tag `breaking-issue`. 
* [ ]  NO
 
 If yes, provide the breaking commit details / MR here. 
 
### Action taken
What action did you take to avoid this in future?

 Feature matrix document updated
* [ ]  Yes
* [ ]  NO
* [ ]  NA
 
Automation details - Mark `Is Automated` field as (Yes, Manual, Not Applicable) in corresponding JIRA task once the bug is automated. 
* [ ] BUnit, share corresponding MR.
* [ ] E2E or Manual Automation using tester - Make sure all items are automated with priority before release which can be tracked in [automation dashboard](https://syncfusion.atlassian.net/secure/Dashboard.jspa?selectPageId=43396).
 
If the same issue is reproduced in ej2, what will you do?
* [ ]  Resolved. Provide MR link.
* [ ]  NO. Created task to track it. Share task link. 
* [ ]  NA
 
 Is this common issue need to be addressed in the same component or on other components in our platform? 
* [ ]  Yes - Need to check in other components, tag `needs-attention-coreteam` 
* [ ]  No
  
### Output screenshots
Post the output screenshots if an UI is affected or added due to this bug.

### Blazor Checklist
Confirm whether this feature is ensured in both Blazor Server and WASM
* [ ]  NA
* [ ]  Yes
* [ ]  NO

Is there any new API or existing API name change?
* [ ]  Yes. If yes, Provide API Review task link.
* [ ]  NO
  
Is there any existing behavior change due to this code change?
* [ ]  Yes. Add `breaking-change` label.
* [ ]  NO


Do the code changes cause any memory leak and performance issue? (Test only if you suspect that your code may cause problem)
* [ ]  Yes
* [ ]  NO

### Reviewer Checklist
* [ ]  All provided information are reviewed and ensured.